backend/requirements.txt
# --- Core Application ---
fastapi[all]
uvicorn
google-genai
google-cloud-firestore
google-cloud-storage
firebase-admin
python-dotenv
pydantic

# --- Testing Suite ---
pytest
pytest-mock
httpx
